commit_again

+++ ft_printf.h
@@ -16,9 +16,9 @@ int		print_number(long n);
 
 int		print_unsigned_int(unsigned long n);
 
-int		print_x(unsigned long long i);
+int		print_x(unsigned long i);
 
-int		print_X(unsigned long long i);
+int		print_X(unsigned long i);
 
 char	*ft_itoa(int n);
